// src/types/Character/Feats/FeatFlag.ts
enum FeatFlag {
    combat = 'combat',
    skills = 'skills',
    skbase = 'skbase',
    trade = 'trade',
    tech = 'tech',
    saves = 'saves',
    expertise = 'expertise',
    critical = 'critical',
    mobility = 'mobility',
    unarmed = 'unarmed',
    shooting = 'shooting',
    power = 'power',
    mounted = 'mounted',
    proficiency = 'proficiency',
    armour = 'armour',
    weapons = 'weapons',
    simple = 'simple',
    martial = 'martial',
    exotic = 'exotic',
    shield = 'shield',
    dualWield = 'dualWield',
    vital = 'vital',
    weaponFocus = 'weaponFocus',
    magic = 'magic',
    spellFocus = 'spellFocus',
    itemCreation = 'itemCreation',
    metaMagic = 'metaMagic',
    class = 'class',
    channel = 'channel',
    byWeapon = 'byWeapon',
    noWeapon = 'noWeapon',
    multiple = 'multiple',
}

export { FeatFlag };